+++ src/compas_timber/fabrication/jack_cut.py
@@ -8,12 +8,12 @@ from compas.geometry import Plane
 from compas.geometry import Rotation
 from compas.geometry import Vector
 from compas.geometry import angle_vectors_signed
-from compas.geometry import distance_point_point
+from compas.geometry import dot_vectors
 from compas.geometry import intersection_line_plane
-from compas.geometry import is_point_behind_plane
 from compas.tolerance import TOL
 
 from compas_timber.errors import FeatureApplicationError
+from compas_timber.utils import planar_surface_point_at
 
 from .btlx import BTLxProcessing
 from .btlx import BTLxProcessingParams
@@ -166,12 +166,12 @@ class JackRafterCut(BTLxProcessing):
         start_depth = 0.0
         ref_side = beam.ref_sides[ref_side_index]  # TODO: is this arbitrary?
         ref_edge = Line.from_point_and_vector(ref_side.point, ref_side.xaxis)
-        orientation = cls._calculate_orientation(ref_side, plane)
+        orientation = cls._calculate_orientation(beam, plane)
         point_start_x = intersection_line_plane(ref_edge, plane)
         if point_start_x is None:
             raise ValueError("Plane does not intersect with beam.")
 
-        start_x = distance_point_point(ref_edge.point, point_start_x)
+        start_x = dot_vectors(Vector.from_start_end(ref_edge.point, point_start_x), ref_side.xaxis)
         angle = cls._calculate_angle(ref_side, plane, orientation)
         inclination = cls._calculate_inclination(ref_side, plane, orientation)
         return cls(orientation, start_x, start_y, start_depth, angle, inclination, ref_side_index=ref_side_index, **kwargs)
@@ -198,10 +198,13 @@ class JackRafterCut(BTLxProcessing):
         return cls.from_plane_and_beam(plane, element, **kwargs)
 
     @staticmethod
-    def _calculate_orientation(ref_side, cutting_plane):
+    def _calculate_orientation(beam, cutting_plane):
         # orientation is START if cutting plane normal points towards the start of the beam and END otherwise
         # essentially if the start is being cut or the end
-        if is_point_behind_plane(ref_side.point, cutting_plane):
+        dot = dot_vectors(beam.frame.xaxis, cutting_plane.normal)
+        if TOL.is_zero(dot):
+            raise ValueError("Plane is parallel to beam, no orientation could be identified.")
+        if dot > 0:
             return OrientationType.END
         else:
             return OrientationType.START
@@ -246,8 +249,7 @@ class JackRafterCut(BTLxProcessing):
 
         """
         # type: (Brep, Beam) -> Brep
-        cutting_plane = self.plane_from_params_and_beam(beam)
-        cutting_plane.transform(beam.transformation_to_local())
+        cutting_plane = self.plane_from_params_and_beam(beam).transformed(beam.transformation_to_local())
         try:
             return geometry.trimmed(cutting_plane)
         except BrepTrimmingError:
@@ -277,7 +279,7 @@ class JackRafterCut(BTLxProcessing):
 
         # start with a plane aligned with the ref side but shifted to the start_x of the cut
         ref_side = beam.side_as_surface(self.ref_side_index)
-        p_origin = ref_side.point_at(self.start_x, 0.0)
+        p_origin = planar_surface_point_at(ref_side, self.start_x, 0.0)
         cutting_plane = Frame(p_origin, ref_side.frame.xaxis, ref_side.frame.yaxis)
 
         # normal pointing towards xaxis so just need the delta
@@ -375,7 +377,7 @@ class JackRafterCutProxy(object):
         return self.unproxified()
 
     def __init__(self, plane, beam, ref_side_index=0):
-        self.plane = plane
+        self.plane = plane.transformed(beam.transformation_to_local())
         self.beam = beam
         self.ref_side_index = ref_side_index
         self._processing = None
@@ -389,7 +391,8 @@ class JackRafterCutProxy(object):
 
         """
         if not self._processing:
-            self._processing = JackRafterCut.from_plane_and_beam(self.plane, self.beam, self.ref_side_index)
+            plane = self.plane.transformed(self.beam.modeltransformation)
+            self._processing = JackRafterCut.from_plane_and_beam(plane, self.beam, self.ref_side_index)
         return self._processing
 
     @classmethod
@@ -434,13 +437,11 @@ class JackRafterCutProxy(object):
 
         """
         # type: (Brep, Beam) -> Brep
-        cutting_plane = self.plane
-        cutting_plane = cutting_plane.transformed(beam.transformation_to_local())
         try:
-            return geometry.trimmed(cutting_plane)
+            return geometry.trimmed(self.plane)
         except BrepTrimmingError:
             raise FeatureApplicationError(
-                cutting_plane,
+                self.plane,
                 geometry,
                 "The cutting plane does not intersect with beam geometry.",
             )