+++ src/compas_timber/model/model.py
@@ -13,20 +13,19 @@ from compas_model.models import Model
 from compas_timber.connections import ConnectionSolver
 from compas_timber.connections import JointCandidate
 from compas_timber.connections import JointTopology
+from compas_timber.connections import PanelJoint
 from compas_timber.connections import PlateConnectionSolver
 from compas_timber.connections import PlateJoint
 from compas_timber.connections import PlateJointCandidate
-from compas_timber.connections import WallJoint
 from compas_timber.elements import Beam
 from compas_timber.elements import Fastener
+from compas_timber.elements import Panel
 from compas_timber.elements import Plate
-from compas_timber.elements import Slab
-from compas_timber.elements import Wall
 from compas_timber.errors import BeamJoiningError
 
 
 class TimberModel(Model):
-    """Represents a timber model containing different elements such as walls, beams and joints.
+    """Represents a timber model containing different elements such as panels, beams and joints.
 
     The timber model allows expressing the hierarchy and interactions between the different elements it contains.
 
@@ -40,8 +39,8 @@ class TimberModel(Model):
         A set of all actual joints assigned to this model.
     joint_candidates : set[:class:`~compas_timber.connections.JointCandidate`]
         A set of all joint candidates in the model.
-    walls : Generator[:class:`~compas_timber.elements.Wall`]
-        A Generator object of all walls assigned to this model.
+    panels : Generator[:class:`~compas_timber.elements.Panel`]
+        A Generator object of all panels assigned to this model.
     center_of_mass : :class:`~compas.geometry.Point`
         The calculated center of mass of the model.
     topologies :  list(dict)
@@ -93,14 +92,9 @@ class TimberModel(Model):
         return self.find_all_elements_of_type(Plate)
 
     @property
-    def walls(self):
-        # type: () -> List[Wall]
-        return self.find_all_elements_of_type(Wall)
-
-    @property
-    def slabs(self):
-        # type: () -> List[Slab]
-        return self.find_all_elements_of_type(Slab)
+    def panels(self):
+        # type: () -> List[Panel]
+        return self.find_all_elements_of_type(Panel)
 
     @property
     def fasteners(self):
@@ -431,11 +425,6 @@ class TimberModel(Model):
         for candidate in list(self.joint_candidates):
             self.remove_joint_candidate(candidate)
 
-        # Clear existing joints (except WallJoints)
-        for joint in list(self.joints):
-            if not isinstance(joint, WallJoint):
-                self.remove_joint(joint)
-
         max_distance = max_distance or TOL.relative
         beams = self.beams
         solver = ConnectionSolver()
@@ -483,49 +472,32 @@ class TimberModel(Model):
             candidate = PlateJointCandidate(result.plate_a, result.plate_b, **kwargs)
             self.add_joint_candidate(candidate)
 
-    def connect_adjacent_walls(self, max_distance=None):
-        """Connects adjacent walls in the model.
+    def connect_adjacent_panels(self, max_distance=None):
+        """Connects adjacent plates in the model.
 
         Parameters
         ----------
         max_distance : float, optional
-            The maximum distance between walls to consider them adjacent. Default is 0.0.
-
+            The maximum distance between plates to consider them adjacent. Default is 0.0.
         """
-        self._clear_wall_joints()
-
-        walls = self.walls
-
-        if not walls:
-            return
-
-        if max_distance is None:
-            max_distance = max(wall.thickness for wall in walls)
-
-        solver = ConnectionSolver()
-        pairs = solver.find_intersecting_pairs(walls, rtree=True, max_distance=max_distance)
-        for pair in pairs:
-            wall_a, wall_b = pair
-            result = solver.find_wall_wall_topology(wall_a, wall_b, tol=self._tolerance.absolute, max_distance=max_distance)
-
-            topology = result[0]
-
-            unsupported_topos = (JointTopology.TOPO_UNKNOWN, JointTopology.TOPO_I, JointTopology.TOPO_X)
-            if topology in unsupported_topos:
-                continue
-
-            wall_a, wall_b = result[1], result[2]
-
-            assert wall_a and wall_b
-
-            # assume wall_a is the main, unless wall_b is explicitly marked as main
-            # TODO: use the Rule system? this isn't good enough, a wall can totally be main and cross at the same time (in two different interactions)
-            if wall_b.attributes.get("role", "cross") == "main":
-                WallJoint.create(self, wall_b, wall_a, topology=topology)
-            else:
-                WallJoint.create(self, wall_a, wall_b, topology=topology)
-
-    def _clear_wall_joints(self):
         for joint in self.joints:
-            if isinstance(joint, WallJoint):
-                self.remove_joint(joint)
+            if isinstance(joint, PanelJoint):
+                self.remove_joint(joint)  # TODO do we want to remove plate joints?
+
+        max_distance = max_distance or TOL.absolute
+        panels = self.panels
+        solver = PlateConnectionSolver()
+        pairs = solver.find_intersecting_pairs(panels, rtree=True, max_distance=max_distance)
+        for pair in pairs:
+            panel_a, panel_b = pair
+            result = solver.find_topology(panel_a, panel_b, tol=TOL.relative, max_distance=max_distance)
+
+            if result.topology is JointTopology.TOPO_UNKNOWN:
+                continue
+            kwargs = {"topology": result.topology, "a_segment_index": result.a_segment_index, "distance": result.distance, "location": result.location}
+
+            if result.topology == JointTopology.TOPO_EDGE_EDGE:
+                kwargs["b_segment_index"] = result.b_segment_index
+
+            candidate = PlateJointCandidate(result.plate_a, result.plate_b, **kwargs)
+            self.add_joint_candidate(candidate)